[consensus] weighted voting
Currently all validators have 1 vote in consensus. In order to support
systems that have weighted voting (e.g. proof-of-stake), Libra Core
must support weighted voting. The main change is to modify
ValidatorVerifier to support weighted voting quorums.

* Unified config access for ValidatorVerifier to get_validator_validator() from multiple locations to a single access point
* Consensus voting is handled directly by ValidatorVerifier rather than passing quorum sizes explicitly
* Added random_validator_verifier to simplify testing that requires ValidatorVerifier and reduce boilerplate
* Added weighted voting counting to BlockTree#insert_vote(), PacemakerTimeoutCertificate#verify(), PacemakerTimeoutManager#generate_timeout_certificate
* Added new tests in ValidatorVerifier#test_unequal_vote_quorum_validators() for weighted voting

Future work
* Add for support configuration that uses voting weights other than 1
* Add more support for weighted voting counting in ValidatorVerifier rather than directly counting in consensus

+++ consensus/src/chained_bft/block_storage/mod.rs
@@ -21,9 +21,9 @@ use executor::StateComputeResult;
 /// as the Error part of the result.
 #[derive(Debug, PartialEq)]
 pub enum VoteReceptionResult {
-    /// The vote has been added but QC has not been formed yet. Return the number of votes for
+    /// The vote has been added but QC has not been formed yet. Return the amount of voting power
     /// the given (proposal, execution) pair.
-    VoteAdded(usize),
+    VoteAdded(u64),
     /// The very same vote message has been processed in past.
     DuplicateVote,
     /// This block has been already certified.
